refactor(engine): extract ability-cost payment authority into game/costs.rs
Phase 1 of the cost-payment unification plan (.planning/cost-payment-unification/PLAN.md):
pure code-motion extraction of the L2 activation cost-payment authority out of the
41k-line casting.rs into a dedicated game/costs.rs seam.

Moved verbatim (byte-identical, no behavior change, no renames):
- AbilityCostPaymentOutcome
- combine_remaining_costs
- pause_cost_payment_for_replacement_choice
- pay_ability_cost / pay_ability_cost_for_activation / pay_ability_cost_inner

casting.rs re-exports these via pub use / pub(crate) use shims so all existing
call sites compile unchanged. The activation flow, WaitingFor::PayCost
emission/resume handlers, can_pay_ability_cost_now affordability aggregate, the
cost finder helpers, and the mana planner intentionally stay in casting.rs (they
belong to the L3 activation adapter / mana planner, split in later plan phases).
